+++ packages/react/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="9.8.0"></a>
+# [9.8.0](https://github.com/xing/hops/compare/v9.7.0...v9.8.0) (2018-01-25)
+
+
+### Features
+
+* **hops-react:** add header component to allow setting http headers declaratively ([f6dfc34](https://github.com/xing/hops/commit/f6dfc34))
+
+
+
+
+<a name="9.5.0"></a>
+# [9.5.0](https://github.com/xing/hops/compare/v9.4.3...v9.5.0) (2018-01-09)
+
+
+
+
+**Note:** Version bump only for package hops-react
+
 <a name="9.3.0"></a>
 # [9.3.0](https://github.com/xing/hops/compare/v9.2.0...v9.3.0) (2017-12-14)