+++ packages/spec/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="9.8.0"></a>
+# [9.8.0](https://github.com/xing/hops/compare/v9.7.0...v9.8.0) (2018-01-25)
+
+
+
+
+**Note:** Version bump only for package hops-spec
+
+<a name="9.7.0"></a>
+# [9.7.0](https://github.com/xing/hops/compare/v9.6.1...v9.7.0) (2018-01-22)
+
+
+
+
+**Note:** Version bump only for package hops-spec
+
+<a name="9.6.1"></a>
+## [9.6.1](https://github.com/xing/hops/compare/v9.6.0...v9.6.1) (2018-01-12)
+
+
+
+
+**Note:** Version bump only for package hops-spec
+
+<a name="9.6.0"></a>
+# [9.6.0](https://github.com/xing/hops/compare/v9.5.0...v9.6.0) (2018-01-09)
+
+
+
+
+**Note:** Version bump only for package hops-spec
+
+<a name="9.5.0"></a>
+# [9.5.0](https://github.com/xing/hops/compare/v9.4.3...v9.5.0) (2018-01-09)
+
+
+
+
+**Note:** Version bump only for package hops-spec
+
 <a name="9.4.3"></a>
 ## [9.4.3](https://github.com/xing/hops/compare/v9.4.2...v9.4.3) (2018-01-08)